% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_message.R
\name{create_message}
\alias{create_message}
\title{Send a message}
\usage{
create_message(
  to,
  text = NULL,
  from = NULL,
  media_urls = NULL,
  messaging_profile_id = NULL,
  subject = NULL,
  auto_detect = FALSE,
  type = NULL,
  use_profile_webhooks = TRUE,
  webhook_failover_url = NULL,
  webhook_url = NULL
)
}
\arguments{
\item{to}{string (address) Receiving address (+18665552367 formatted phone number or short code).
Example: "+18665552367"}

\item{text}{string Message body (i.e., content) as a non-empty string. **Required for SMS**}

\item{from}{string (address) Sending address (+18665552367 formatted phone number, alphanumeric
sender ID, or short code). **Required if sending with a phone number, short code, or
alphanumeric sender ID.**}

\item{media_urls}{array of string A list of media URLs. The total media size must be less than
1 MB. **Required for MMS**}

\item{messaging_profile_id}{string Unique identifier for a messaging profile. **Required if
sending via number pool or with an alphanumeric sender ID.**}

\item{subject}{string Subject of multimedia message.}

\item{auto_detect}{boolean Automatically detect if an SMS message is unusually long and exceeds
a recommended limit of message parts.}

\item{type}{string The protocol for sending the message, either SMS or MMS.
Options: ["SMS", "MMS"]}

\item{use_profile_webhooks}{boolean If the profile this number is associated with has webhooks,
use them for delivery notifications. If webhooks are also specified on the message itself,
they will be attempted first, then those on the profile.}

\item{webhook_failover_url}{string (url) The failover URL where webhooks related to this message
will be sent if sending to the primary URL fails.}

\item{webhook_url}{string (url) The URL where webhooks related to this message will be sent.}
}
\description{
Send a message with a Phone Number, Alphanumeric Sender ID, Short Code or Number Pool.
This endpoint allows you to send a message with any messaging resource. Current messaging
resources include: long-code, short-code, number-pool, and alphanumeric-sender-id.
}
