% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_short_code_message.R
\name{create_short_code_message}
\alias{create_short_code_message}
\title{Send a short code message}
\usage{
create_short_code_message(
  from,
  to,
  text = NULL,
  media_urls = NULL,
  subject = NULL,
  auto_detect = FALSE,
  type = NULL,
  use_profile_webhooks = TRUE,
  webhook_failover_url = NULL,
  webhook_url = NULL
)
}
\arguments{
\item{from}{string (address) Phone number, in +E.164 format, used to send the message.}

\item{to}{string (address) Receiving address (+E.164 formatted phone number or short code).
Example: "+E.164"}

\item{text}{string Message body (i.e., content) as a non-empty string. **Required for SMS**}

\item{media_urls}{array of string A list of media URLs. The total media size must be less than
1 MB. **Required for MMS**}

\item{subject}{string Subject of multimedia message.}

\item{auto_detect}{boolean Automatically detect if an SMS message is unusually long and exceeds
a recommended limit of message parts.}

\item{type}{string The protocol for sending the message, either SMS or MMS.
Options: ["SMS", "MMS"]}

\item{use_profile_webhooks}{boolean If the profile this number is associated with has webhooks,
use them for delivery notifications. If webhooks are also specified on the message itself,
they will be attempted first, then those on the profile.}

\item{webhook_failover_url}{string (url) The failover URL where webhooks related to this message
will be sent if sending to the primary URL fails.}

\item{webhook_url}{string (url) The URL where webhooks related to this message will be sent.}
}
\description{
Send a short code message.
}
