% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paginated_mdrs.R
\name{get_paginated_mdrs}
\alias{get_paginated_mdrs}
\title{Fetch messaging detail records}
\usage{
get_paginated_mdrs(
  start_date,
  end_date,
  page_number = 1,
  page_size = 20,
  sort = c("created_at"),
  id = NULL,
  direction = NULL,
  outbound_profile_id = NULL,
  cld = NULL,
  cli = NULL,
  status = NULL,
  message_type = NULL,
  country_iso = NULL,
  error = NULL,
  normalized_carrier = NULL,
  tag = NULL,
  mcc = NULL,
  mnc = NULL,
  product = NULL
)
}
\arguments{
\item{start_date}{date-time Start date. Example: Sys.time()}

\item{end_date}{date-time End date. Example: Sys.time()}

\item{page_number}{integer (1) Page number.}

\item{page_size}{integer (1 - 50) Size of the page.}

\item{sort}{array of string Field used to order the data. Example: "created_at"}

\item{id}{string Message uuid. Example: "e093fbe0-5bde-11eb-ae93-0242ac130002"}

\item{direction}{string Options: ["INBOUND", "OUTBOUND"]}

\item{outbound_profile_id}{string Configured profile id. New profiles can be created and
configured on Telnyx portal. Example: "30ef55db-c4a2-4c4a-9804-a68077973d07"}

\item{cld}{string Destination number. Example: "+15551237654"}

\item{cli}{string Origination number. Example: "+15551237654"}

\item{status}{string Message status. Options: ["GW_TIMEOUT", "DELIVERED", "DLR_UNCONFIRMED",
"DLR_TIMEOUT", "RECEIVED", "GW_REJECT", "FAILED"]}

\item{message_type}{string type of the message. Default: "text". Options: ["audio", "contacts",
"document", "hsm", "image", "location", "template", "text", "video", "voice", "unknown"]}

\item{country_iso}{string Destination number ISO country code. Example: "US"}

\item{error}{string Reported error code. Example: "40001"}

\item{normalized_carrier}{string Carrier. Example: "Verizon"}

\item{tag}{string Specified tag. Example: "Tag1"}

\item{mcc}{string Mobile country code. Example: "204"}

\item{mnc}{string Mobile network code. Example: "01"}

\item{product}{string Used product. Options: ["LONG_CODE", "SHORT_CODE", "SHORT_CODE_FTEU",
"TOLL_FREE", "ALPHANUMERIC_ID", "RCS"]}
}
\description{
Fetch all Mdr records.
}
