% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_messaging_profile_phone_numbers.R
\name{list_messaging_profile_phone_numbers}
\alias{list_messaging_profile_phone_numbers}
\title{List phone numbers associated with a messaging profile}
\usage{
list_messaging_profile_phone_numbers(id, page_number = 1, page_size = 20)
}
\arguments{
\item{id}{string (uuid) The id of the messaging profile to retrieve.}

\item{page_number}{integer (1) The page number to load.}

\item{page_size}{integer (1 - 250) The size of the page.}
}
\description{
List phone numbers associated with a messaging profile.
}
