% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_messaging_profiles.R
\name{list_messaging_profiles}
\alias{list_messaging_profiles}
\title{List messaging profiles}
\usage{
list_messaging_profiles(page_number = 1, page_size = 20)
}
\arguments{
\item{page_number}{integer (1) The page number to load.}

\item{page_size}{integer (1 - 250) The size of the page.}
}
\description{
List messaging profiles.
}
