% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_phone_numbers.R
\name{list_phone_numbers}
\alias{list_phone_numbers}
\title{List phone numbers}
\usage{
list_phone_numbers(
  page_number = 1,
  page_size = 20,
  filter_tag = NULL,
  filter_phone_number = NULL,
  filter_status = NULL,
  filter_voice_connection_name_contains = NULL,
  filter_voice_connection_name_starts_with = NULL,
  filter_voice_connection_name_ends_with = NULL,
  filter_voice_connection_name_eq = NULL,
  filter_usage_payment_method = NULL,
  filter_billing_group_id = NULL,
  filter_emergency_address_id = NULL,
  filter_customer_reference = NULL,
  sort = NULL
)
}
\arguments{
\item{page_number}{integer (1) The page number to load.}

\item{page_size}{integer (1 - 250) The size of the page.}

\item{filter_tag}{string Filter by phone number tags.}

\item{filter_phone_number}{string Filter by phone number. Requires at least three digits.
Non-numerical characters will result in no values being returned.}

\item{filter_status}{string Filter by phone number status. Options: ["purchase_pending",
"purchase_failed", "port_pending", "active", "deleted", "port_failed", "emergency_only",
"ported_out", "port_out_pending"]}

\item{filter_voice_connection_name_contains}{string Filter contains connection name. Requires at
least three characters. Example: "test"}

\item{filter_voice_connection_name_starts_with}{string Filter starts with connection name.
Requires at least three characters. Example: "test"}

\item{filter_voice_connection_name_ends_with}{string Filter ends with connection name. Requires
at least three characters. Example: "test"}

\item{filter_voice_connection_name_eq}{string Filter by connection name. Example: "test"}

\item{filter_usage_payment_method}{string Filter by usage_payment_method.
Options: ["pay-per-minute", "channel"]}

\item{filter_billing_group_id}{string Filter by the billing_group_id associated with phone
numbers. To filter to only phone numbers that have no billing group associated them, set the
value of this filter to the string 'null'. Example: "62e4bf2e-c278-4282-b524-488d9c9c43b2"}

\item{filter_emergency_address_id}{string (int64) Filter by the emergency_address_id associated
with phone numbers. To filter only phone numbers that have no emergency address associated
with them, set the value of this filter to the string 'null'. Example: "9102160989215728032"}

\item{filter_customer_reference}{string Filter numbers via the customer_reference set.}

\item{sort}{string Specifies the sort order for results. If not given, results are sorted by
created_at in descending order. Options: ["purchased_at", "phone_number", "connection_name",
"usage_payment_method"]}
}
\description{
List phone numbers.
}
