% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_message.R
\name{retrieve_message}
\alias{retrieve_message}
\title{Retrieve a message}
\usage{
retrieve_message(id)
}
\arguments{
\item{id}{string (uuid) The id of the message.}
}
\description{
Note: This API endpoint can only retrieve messages that are no older than 10 days since their
creation. If you require messages older than this, please generate an MDR report.
}
