% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telnyxr-package.R
\docType{package}
\name{telnyxr-package}
\alias{telnyxr}
\alias{telnyxr-package}
\title{`{telnyxr}`: An Interface to the Telnyx API for R}
\description{
`{telnyxr}` provides users a range of functions designed to communicate with Telnyx API from R.
}
\author{
\strong{Maintainer}: Who Wrote It \email{who@wrote.it}

}
