% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telnyxr_auth.R
\name{telnyxr_auth}
\alias{telnyxr_auth}
\title{Set Telnyx's Credentials}
\usage{
telnyxr_auth(token = Sys.getenv("TELNYX_TOKEN"))
}
\arguments{
\item{token}{Telnyx's token.}
}
\description{
Set Telnyx's Credentials
}
